+++ heat/core/tests/test_manipulations.py
@@ -725,6 +725,61 @@ class TestManipulations(TestCase):
             numpy_args={"axis1": 0, "axis2": 1},
         )
 
+    def test_dsplit(self):
+        # for further testing, see test_split
+        data_ht = ht.arange(24).reshape((2, 3, 4))
+        data_np = data_ht.numpy()
+
+        # indices_or_sections = int
+        result = ht.dsplit(data_ht, 2)
+        comparison = np.dsplit(data_np, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = tuple
+        result = ht.dsplit(data_ht, (0, 1))
+        comparison = np.dsplit(data_np, (0, 1))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = list
+        result = ht.dsplit(data_ht, [0, 1])
+        comparison = np.dsplit(data_np, [0, 1])
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = undistributed DNDarray
+        result = ht.dsplit(data_ht, ht.array([0, 1]))
+        comparison = np.dsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = distributed DNDarray
+        result = ht.dsplit(data_ht, ht.array([0, 1], split=0))
+        comparison = np.dsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
     def test_expand_dims(self):
         # vector data
         a = ht.arange(10)
@@ -950,6 +1005,115 @@ class TestManipulations(TestCase):
             dtype=ht.int8,
         )
         self.assertTrue(ht.equal(ht.resplit(ht.flipud(c), 0), r_c))
+
+    def test_hsplit(self):
+        # for further testing, see test_split
+        # 1-dimensional array (as forbidden in split)
+        data_ht = ht.arange(24)
+        data_np = data_ht.numpy()
+
+        # indices_or_sections = int
+        result = ht.hsplit(data_ht, 2)
+        comparison = np.hsplit(data_np, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = tuple
+        result = ht.hsplit(data_ht, (0, 1))
+        comparison = np.hsplit(data_np, (0, 1))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = list
+        result = ht.hsplit(data_ht, [0, 1])
+        comparison = np.hsplit(data_np, [0, 1])
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = undistributed DNDarray
+        result = ht.hsplit(data_ht, ht.array([0, 1]))
+        comparison = np.hsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = distributed DNDarray
+        result = ht.hsplit(data_ht, ht.array([0, 1], split=0))
+        comparison = np.hsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        data_ht = ht.arange(24).reshape((2, 4, 3))
+        data_np = data_ht.numpy()
+
+        # indices_or_sections = int
+        result = ht.hsplit(data_ht, 2)
+        comparison = np.hsplit(data_np, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = tuple
+        result = ht.hsplit(data_ht, (0, 1))
+        comparison = np.hsplit(data_np, (0, 1))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = list
+        result = ht.hsplit(data_ht, [0, 1])
+        comparison = np.hsplit(data_np, [0, 1])
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = undistributed DNDarray
+        result = ht.hsplit(data_ht, ht.array([0, 1]))
+        comparison = np.hsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = distributed DNDarray
+        result = ht.hsplit(data_ht, ht.array([0, 1], split=0))
+        comparison = np.hsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
 
     def test_hstack(self):
         # cases to test:
@@ -2167,6 +2331,220 @@ class TestManipulations(TestCase):
                 if rank == i:
                     self.assertTrue(torch.lt(result.larray[idx], result.larray[idx + 1]).all())
 
+    def test_split(self):
+        # ====================================
+        # UNDISTRIBUTED CASE
+        # ====================================
+        # axis = 0
+        # ====================================
+        data_ht = ht.arange(24).reshape((2, 3, 4))
+        data_np = data_ht.numpy()
+
+        # indices_or_sections = int
+        result = ht.split(data_ht, 2)
+        comparison = np.split(data_np, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = tuple
+        result = ht.split(data_ht, (0, 1))
+        comparison = np.split(data_np, (0, 1))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = list
+        result = ht.split(data_ht, [0, 1])
+        comparison = np.split(data_np, [0, 1])
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = undistributed DNDarray
+        result = ht.split(data_ht, ht.array([0, 1]))
+        comparison = np.split(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = distributed DNDarray
+        result = ht.split(data_ht, ht.array([0, 1], split=0))
+        comparison = np.split(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # ====================================
+        # axis != 0 (2 in this case)
+        # ====================================
+        # indices_or_sections = int
+        result = ht.split(data_ht, 2, 2)
+        comparison = np.split(data_np, 2, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = tuple
+        result = ht.split(data_ht, (0, 1))
+        comparison = np.split(data_np, (0, 1))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # exceptions
+        with self.assertRaises(TypeError):
+            ht.split([1, 2, 3, 4], 2)
+        with self.assertRaises(TypeError):
+            ht.split(data_ht, "2")
+        with self.assertRaises(TypeError):
+            ht.split(data_ht, 2, "0")
+        with self.assertRaises(ValueError):
+            ht.split(data_ht, 2, -1)
+        with self.assertRaises(ValueError):
+            ht.split(data_ht, 2, 3)
+        with self.assertRaises(ValueError):
+            ht.split(data_ht, 5)
+        with self.assertRaises(ValueError):
+            ht.split(data_ht, [[0, 1]])
+
+        # ====================================
+        # DISTRIBUTED CASE
+        # ====================================
+        # axis == ary.split
+        # ====================================
+        data_ht = ht.arange(120, split=0).reshape((4, 5, 6))
+        data_np = data_ht.numpy()
+
+        # indices = int
+        result = ht.split(data_ht, 2)
+        comparison = np.split(data_np, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assertTrue((ht.array(comparison[i]) == result[i]).all())
+
+        # larger example
+        data_ht_large = ht.arange(160, split=0).reshape((8, 5, 4))
+        data_np_large = data_ht_large.numpy()
+
+        # indices = int
+        result = ht.split(data_ht_large, 2)
+        comparison = np.split(data_np_large, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assertTrue((ht.array(comparison[i]) == result[i]).all())
+
+        # indices_or_sections = tuple
+        result = ht.split(data_ht, (1, 3, 5))
+        comparison = np.split(data_np, (1, 3, 5))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = list
+        result = ht.split(data_ht, [1, 3, 5])
+        comparison = np.split(data_np, [1, 3, 5])
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = undistributed DNDarray
+        result = ht.split(data_ht, ht.array([1, 3, 5]))
+        comparison = np.split(data_np, np.array([1, 3, 5]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = distributed DNDarray
+        result = ht.split(data_ht, ht.array([1, 3, 5], split=0))
+        comparison = np.split(data_np, np.array([1, 3, 5]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # ====================================
+        # axis != ary.split
+        # ====================================
+        # indices_or_sections = int
+        result = ht.split(data_ht, 2, 2)
+        comparison = np.split(data_np, 2, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = list
+        result = ht.split(data_ht, [3, 4, 6], 2)
+        comparison = np.split(data_np, [3, 4, 6], 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = undistributed DNDarray
+        result = ht.split(data_ht, ht.array([3, 4, 6]), 2)
+        comparison = np.split(data_np, np.array([3, 4, 6]), 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = distributed DNDarray
+        indices = ht.array([3, 4, 6], split=0)
+        result = ht.split(data_ht, indices, 2)
+        comparison = np.split(data_np, np.array([3, 4, 6]), 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
     def test_resplit(self):
         if ht.MPI_WORLD.size > 1:
             # resplitting with same axis, should leave everything unchanged
@@ -2599,6 +2977,61 @@ class TestManipulations(TestCase):
         data_split_zero = ht.array(torch_array, split=0)
         res, inv = ht.unique(data_split_zero, return_inverse=True, sorted=True)
         self.assertTrue(torch.equal(inv, exp_inv.to(dtype=inv.dtype)))
+
+    def test_vsplit(self):
+        # for further testing, see test_split
+        data_ht = ht.arange(24).reshape((4, 3, 2))
+        data_np = data_ht.numpy()
+
+        # indices_or_sections = int
+        result = ht.vsplit(data_ht, 2)
+        comparison = np.vsplit(data_np, 2)
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = tuple
+        result = ht.vsplit(data_ht, (0, 1))
+        comparison = np.vsplit(data_np, (0, 1))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = list
+        result = ht.vsplit(data_ht, [0, 1])
+        comparison = np.vsplit(data_np, [0, 1])
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = undistributed DNDarray
+        result = ht.vsplit(data_ht, ht.array([0, 1]))
+        comparison = np.vsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
+
+        # indices_or_sections = distributed DNDarray
+        result = ht.vsplit(data_ht, ht.array([0, 1], split=0))
+        comparison = np.vsplit(data_np, np.array([0, 1]))
+
+        self.assertTrue(len(result) == len(comparison))
+
+        for i in range(len(result)):
+            self.assertIsInstance(result[i], ht.DNDarray)
+            self.assert_array_equal(result[i], comparison[i])
 
     def test_vstack(self):
         # cases to test: